@@ -1,5 +1,5 @@
 {
-  "name": "ordercloud-base-buyer",
+  "name": "ngx-seller",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -13,51 +13,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
-    "@fortawesome/angular-fontawesome": "0.1.0-10",
-    "@fortawesome/fontawesome-svg-core": "^1.2.0-10",
-    "@fortawesome/free-brands-svg-icons": "^5.1.0-4",
-    "@fortawesome/free-regular-svg-icons": "^5.1.0-4",
-    "@fortawesome/free-solid-svg-icons": "^5.1.0-4",
-    "@ng-bootstrap/ng-bootstrap": "^2.0.0-alpha.0",
-    "@ngx-progressbar/core": "^5.0.1",
-    "@ngx-progressbar/http": "^5.0.1",
-    "@ordercloud/angular-sdk": "0.0.1-15",
+    "@angular/animations": "^7.0.4",
+    "@angular/common": "^7.0.4",
+    "@angular/compiler": "^7.0.4",
+    "@angular/core": "^7.0.4",
+    "@angular/forms": "^7.0.4",
+    "@angular/http": "^7.0.4",
+    "@angular/platform-browser": "^7.0.4",
+    "@angular/platform-browser-dynamic": "^7.0.4",
+    "@angular/platform-server": "^7.0.4",
+    "@angular/router": "^7.0.4",
+    "@nguniversal/express-engine": "^7.0.2",
+    "@fortawesome/angular-fontawesome": "^0.3.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.8",
+    "@fortawesome/free-brands-svg-icons": "^5.5.0",
+    "@fortawesome/free-regular-svg-icons": "^5.5.0",
+    "@fortawesome/free-solid-svg-icons": "^5.5.0",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@ngx-progressbar/core": "^5.2.1",
+    "@ngx-progressbar/http": "^5.2.1",
+    "@ordercloud/angular-sdk": "0.0.1-26",
     "@types/lodash": "^4.14.91",
     "@types/moment": "^2.13.0",
     "angular-fontawesome": "^1.0.0",
     "angular-tree-component": "^7.2.1",
     "angular2-image-zoom": "^1.2.1",
     "angular2-text-mask": "^9.0.0",
-    "bootstrap": "^4.0.0-beta.2",
+    "bootstrap": "^4.0.0",
     "core-js": "^2.5.7",
     "jquery": "3",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.4",
     "moment": "^2.20.0",
-    "ngx-cookie": "^4.0.2",
-    "ngx-toastr": "^8.1.1",
+    "ngx-cookie": "^4.1.2",
+    "ngx-toastr": "^9.1.1",
     "popper.js": "^1.14.3",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.3.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "^7.0.6",
+    "@angular/compiler-cli": "^7.0.4",
+    "@angular/language-service": "^7.0.4",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "js-beautify": "^1.7.5",
@@ -72,6 +74,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "~3.1.6"
   }
 }
