@@ -2,7 +2,9 @@ import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { ListShipment, OcMeService } from '@ordercloud/angular-sdk';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root',
+})
 export class ShipmentsResolve implements Resolve<ListShipment> {
   constructor(private ocMeService: OcMeService) {}
 
