@@ -4,7 +4,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { ProfileComponent } from '@app-buyer/profile/containers/profile/profile.component';
 import { AddressListComponent } from '@app-buyer/profile/containers/address-list/address-list.component';
-import { RegisterComponent } from '@app-buyer/shared/containers/register/register.component';
+import { MeUpdateComponent } from '@app-buyer/profile/containers/me-update/me-update.component';
 import { PaymentListComponent } from '@app-buyer/profile/containers/payment-list/payment-list.component';
 
 const routes: Routes = [
@@ -13,11 +13,7 @@ const routes: Routes = [
     component: ProfileComponent,
     children: [
       { path: '', redirectTo: 'details' },
-      {
-        path: 'details',
-        component: RegisterComponent,
-        data: { shouldAllowUpdate: true },
-      },
+      { path: 'details', component: MeUpdateComponent },
       { path: 'addresses', component: AddressListComponent },
       { path: 'payment-methods', component: PaymentListComponent },
       { path: 'orders', loadChildren: '../order/order.module#OrderModule' },
