@@ -1,10 +1,10 @@
-import { TestBed, async, inject } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 
 import { HasTokenGuard } from '@app-buyer/shared/guards/has-token/has-token.guard';
 import { OcTokenService } from '@ordercloud/angular-sdk';
 import { Router } from '@angular/router';
 import { AppAuthService } from '@app-buyer/auth';
-import { of, BehaviorSubject } from 'rxjs';
+import { of } from 'rxjs';
 import { applicationConfiguration } from '@app-buyer/config/app.config';
 import { AppStateService } from '@app-buyer/shared/services/app-state/app-state.service';
 
@@ -42,7 +42,6 @@ describe('HasTokenGuard', () => {
     TestBed.configureTestingModule({
       imports: [],
       providers: [
-        HasTokenGuard,
         { provide: applicationConfiguration, useValue: appConfig },
         { provide: AppAuthService, useValue: appAuthService },
         { provide: Router, useValue: router },
