@@ -6,7 +6,9 @@ import { OcMeService, BuyerProduct, LineItem } from '@ordercloud/angular-sdk';
 import { forEach as _forEach, differenceBy as _differenceBy } from 'lodash';
 import { AppLineItemService } from '@app-buyer/shared/services/line-item/line-item.service';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root',
+})
 export class AppReorderService {
   constructor(
     private appLineItemService: AppLineItemService,
