@@ -13,21 +13,14 @@ import { TreeModule } from 'angular-tree-component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 import { SharedRoutingModule } from '@app-seller/shared/shared-routing.module';
-import { HasTokenGuard } from '@app-seller/shared/guards/has-token/has-token.guard';
-import { AppErrorHandler } from '@app-seller/config/error-handling.config';
-import { AppStateService } from '@app-seller/shared/services/app-state/app-state.service';
 import { SearchComponent } from '@app-seller/shared/components/search/search.component';
 import { SortColumnComponent } from '@app-seller/shared/components/sort-column/sort-column.component';
-import { ModalService } from '@app-seller/shared/services/modal/modal.service';
 import { ModalComponent } from '@app-seller/shared/components/modal/modal.component';
-import { AppFormErrorService } from '@app-seller/shared/services/form-error/form-error.service';
 import { CarouselSlideDisplayComponent } from '@app-seller/shared/components/carousel-slide-display/carousel-slide-display.component';
 import { GenericBrowseComponent } from '@app-seller/shared/components/generic-browse/generic-browse.component';
 import { UserTableComponent } from '@app-seller/shared/containers/user-table/user-table.component';
-import { RegexService } from '@app-seller/shared/services/regex/regex.service';
 import { UserFormComponent } from '@app-seller/shared/components/user-form/user-form.component';
 import { AddressTableComponent } from './containers/address-table/address-table.component';
-import { AppGeographyService } from '@app-seller/shared/services/geography/geography.service';
 import { AddressFormComponent } from '@app-seller/shared/components/address-form/address-form.component';
 import { CategoryTableComponent } from './containers/category-table/category-table.component';
 import { CategoryFormComponent } from './components/category-form/category-form.component';
@@ -94,30 +87,5 @@ import { ProductFormComponent } from '@app-seller/shared/components/products-for
     ProductTableComponent,
     ProductFormComponent,
   ],
-
-  /**
-   * DO NOT define providers here
-   * define providers in the static forRoot below to ensure
-   * lazy-loaded modules define services as singletons
-   * https://angular-2-training-book.rangle.io/handout/modules/shared-di-tree.html
-   */
-  providers: [
-    HasTokenGuard,
-    AppStateService,
-    ModalService,
-    AppFormErrorService,
-    AppGeographyService,
-    RegexService,
-  ],
 })
-export class SharedModule {
-  static forRoot(): ModuleWithProviders {
-    return {
-      ngModule: SharedModule,
-      providers: [
-        AppErrorHandler,
-        { provide: ErrorHandler, useClass: AppErrorHandler },
-      ],
-    };
-  }
-}
+export class SharedModule {}
